drop state transition (moved to anodar/state_transition)
